Created fullstack tests
Tests using postman/newman on a temporary db and running instance of the
server. Tests not all passing yet (there's a troubling issue with
deleting a vtable columns hanging). Reverted changes to the access
function that was made a couple days ago- turns out that change was
completely broken and I didn't realize it because I didn't have tests :)